Phase 2 step 3: Extract watchlist_io + csv_import + agent_io
